Register "Edit->Metadata"; break EditMenus -> Export ...
... Also Export shifts some code lower into TagsEditor
