+++ src/TagsEditor.cpp
@@ -10,6 +10,7 @@
 
 #include "TagsEditor.h"
 
+#include "ProjectWindows.h"
 #include "SelectFile.h"
 #include "ShuttleGui.h"
 #include "widgets/AudacityMessageBox.h"
@@ -941,4 +942,61 @@ bool TagsEditorDialog::IsWindowRectValid(const wxRect *windowRect) const
    }
 
    return true;
+}
+
+#include "Project.h"
+#include "ProjectHistory.h"
+#include "ProjectSettings.h"
+#include <wx/frame.h>
+
+bool TagsEditorDialog::DoEditMetadata(AudacityProject &project,
+   const TranslatableString &title,
+   const TranslatableString &shortUndoDescription, bool force)
+{
+   auto &settings = ProjectSettings::Get( project );
+   auto &tags = Tags::Get( project );
+
+   // Back up my tags
+   // Tags (artist name, song properties, MP3 ID3 info, etc.)
+   // The structure may be shared with undo history entries
+   // To keep undo working correctly, always replace this with a NEW duplicate
+   // BEFORE doing any editing of it!
+   auto newTags = tags.Duplicate();
+
+   if (TagsEditorDialog::ShowEditDialog(
+      *newTags, &GetProjectFrame( project ), title, force)) {
+      if (tags != *newTags) {
+         // Commit the change to project state only now.
+         Tags::Set( project, newTags );
+         ProjectHistory::Get( project ).PushState( title, shortUndoDescription);
+      }
+      bool bShowInFuture;
+      gPrefs->Read(wxT("/AudioFiles/ShowId3Dialog"), &bShowInFuture, true);
+      settings.SetShowId3Dialog( bShowInFuture );
+      return true;
+   }
+
+   return false;
+}
+
+// Attach menu item
+#include "commands/CommandContext.h"
+#include "commands/CommandManager.h"
+#include "CommonCommandFlags.h"
+
+namespace {
+void OnEditMetadata(const CommandContext &context)
+{
+   auto &project = context.project;
+   (void)TagsEditorDialog::DoEditMetadata( project,
+      XO("Edit Metadata Tags"), XO("Metadata Tags"), true);
+}
+
+using namespace MenuTable;
+
+AttachedItem sAttachment{
+   wxT("Edit/Other"),
+   Command( wxT("EditMetaData"), XXO("&Metadata"), OnEditMetadata,
+      AudioIONotBusyFlag() )
+};
 }
